+++ rust/operator-binary/src/airflow_controller.rs
@@ -19,24 +19,30 @@ use stackable_operator::{
         configmap::ConfigMapBuilder,
         meta::ObjectMetaBuilder,
         pod::{
-            PodBuilder, container::ContainerBuilder, resources::ResourceRequirementsBuilder,
-            security::PodSecurityContextBuilder, volume::VolumeBuilder,
+            PodBuilder,
+            container::ContainerBuilder,
+            resources::ResourceRequirementsBuilder,
+            security::PodSecurityContextBuilder,
+            volume::{
+                ListenerOperatorVolumeSourceBuilder, ListenerOperatorVolumeSourceBuilderError,
+                ListenerReference, VolumeBuilder,
+            },
         },
     },
     cluster_resources::{ClusterResourceApplyStrategy, ClusterResources},
-    commons::{
-        authentication::{AuthenticationClass, ldap},
-        product_image_selection::ResolvedProductImage,
-        rbac::build_rbac_resources,
-    },
+    commons::{product_image_selection::ResolvedProductImage, rbac::build_rbac_resources},
     config::fragment::ValidationError,
+    crd::{
+        authentication::{core as auth_core, ldap},
+        git_sync, listener,
+    },
     k8s_openapi::{
         self, DeepMerge,
         api::{
             apps::v1::{StatefulSet, StatefulSetSpec},
             core::v1::{
-                ConfigMap, EmptyDirVolumeSource, EnvVar, PodTemplateSpec, Probe, Service,
-                ServiceAccount, ServicePort, ServiceSpec, TCPSocketAction, VolumeMount,
+                ConfigMap, PersistentVolumeClaim, PodTemplateSpec, Probe, Service, ServiceAccount,
+                ServicePort, ServiceSpec, TCPSocketAction,
             },
         },
         apimachinery::pkg::{apis::meta::v1::LabelSelector, util::intstr::IntOrString},
@@ -47,11 +53,12 @@ use stackable_operator::{
         core::{DeserializeGuard, error_boundary},
         runtime::{controller::Action, reflector::ObjectRef},
     },
-    kvp::{Label, LabelError, Labels},
+    kvp::{Annotation, Label, LabelError, Labels},
     logging::controller::ReconcilerError,
     product_config_utils::{
-        CONFIG_OVERRIDE_FILE_FOOTER_KEY, CONFIG_OVERRIDE_FILE_HEADER_KEY,
-        transform_all_roles_to_config, validate_all_roles_and_groups_config,
+        CONFIG_OVERRIDE_FILE_FOOTER_KEY, CONFIG_OVERRIDE_FILE_HEADER_KEY, env_vars_from,
+        env_vars_from_rolegroup_config, transform_all_roles_to_config,
+        validate_all_roles_and_groups_config,
     },
     product_logging::{
         self,
@@ -76,19 +83,17 @@ use crate::{
     crd::{
         self, AIRFLOW_CONFIG_FILENAME, AIRFLOW_UID, APP_NAME, AirflowClusterStatus, AirflowConfig,
         AirflowConfigOptions, AirflowExecutor, AirflowRole, CONFIG_PATH, Container, ExecutorConfig,
-        ExecutorConfigFragment, LOG_CONFIG_DIR, OPERATOR_NAME, STACKABLE_LOG_DIR,
-        TEMPLATE_CONFIGMAP_NAME, TEMPLATE_LOCATION, TEMPLATE_NAME, TEMPLATE_VOLUME_NAME,
+        ExecutorConfigFragment, HTTP_PORT, HTTP_PORT_NAME, LISTENER_VOLUME_DIR,
+        LISTENER_VOLUME_NAME, LOG_CONFIG_DIR, METRICS_PORT, METRICS_PORT_NAME, OPERATOR_NAME,
+        STACKABLE_LOG_DIR, TEMPLATE_CONFIGMAP_NAME, TEMPLATE_LOCATION, TEMPLATE_NAME,
+        TEMPLATE_VOLUME_NAME,
         authentication::{
             AirflowAuthenticationClassResolved, AirflowClientAuthenticationDetailsResolved,
         },
         authorization::AirflowAuthorizationResolved,
-        build_recommended_labels,
-        git_sync::{GIT_SYNC_CONTENT, GIT_SYNC_NAME, GIT_SYNC_ROOT, GitSync},
-        v1alpha1,
+        build_recommended_labels, v1alpha1,
     },
-    env_vars::{
-        self, build_airflow_template_envs, build_gitsync_statefulset_envs, build_gitsync_template,
-    },
+    env_vars::{self, build_airflow_template_envs},
     operations::{
         graceful_shutdown::{
             add_airflow_graceful_shutdown_config, add_executor_graceful_shutdown_config,
@@ -102,9 +107,6 @@ pub const AIRFLOW_CONTROLLER_NAME: &str = "airflowcluster";
 pub const DOCKER_IMAGE_BASE_NAME: &str = "airflow";
 pub const AIRFLOW_FULL_CONTROLLER_NAME: &str =
     concatcp!(AIRFLOW_CONTROLLER_NAME, '.', OPERATOR_NAME);
-
-const METRICS_PORT_NAME: &str = "metrics";
-const METRICS_PORT: i32 = 9102;
 
 pub struct Ctx {
     pub client: stackable_operator::client::Client,
@@ -176,7 +178,7 @@ pub enum Error {
     #[snafu(display("failed to retrieve AuthenticationClass {authentication_class}"))]
     AuthenticationClassRetrieval {
         source: stackable_operator::cluster_resources::Error,
-        authentication_class: ObjectRef<AuthenticationClass>,
+        authentication_class: ObjectRef<auth_core::v1alpha1::AuthenticationClass>,
     },
 
     #[snafu(display(
@@ -185,7 +187,7 @@ pub enum Error {
     ))]
     AuthenticationClassProviderNotSupported {
         authentication_class_provider: String,
-        authentication_class: ObjectRef<AuthenticationClass>,
+        authentication_class: ObjectRef<auth_core::v1alpha1::AuthenticationClass>,
     },
 
     #[snafu(display("failed to build config file for {rolegroup}"))]
@@ -218,6 +220,9 @@ pub enum Error {
     InvalidContainerName {
         source: stackable_operator::builder::pod::container::Error,
     },
+
+    #[snafu(display("invalid git-sync specification"))]
+    InvalidGitSyncSpec { source: git_sync::v1alpha1::Error },
 
     #[snafu(display("failed to create cluster resources"))]
     CreateClusterResources {
@@ -283,7 +288,7 @@ pub enum Error {
     #[snafu(display(
         "failed to build volume or volume mount spec for the LDAP backend TLS config"
     ))]
-    VolumeAndMounts { source: ldap::Error },
+    VolumeAndMounts { source: ldap::v1alpha1::Error },
 
     #[snafu(display("failed to construct config"))]
     ConstructConfig { source: config::Error },
@@ -305,9 +310,7 @@ pub enum Error {
     },
 
     #[snafu(display("failed to add LDAP Volumes and VolumeMounts"))]
-    AddLdapVolumesAndVolumeMounts {
-        source: stackable_operator::commons::authentication::ldap::Error,
-    },
+    AddLdapVolumesAndVolumeMounts { source: ldap::v1alpha1::Error },
 
     #[snafu(display("failed to add TLS Volumes and VolumeMounts"))]
     AddTlsVolumesAndVolumeMounts {
@@ -321,6 +324,21 @@ pub enum Error {
 
     #[snafu(display("failed to build Statefulset environmental variables"))]
     BuildStatefulsetEnvVars { source: env_vars::Error },
+
+    #[snafu(display("failed to build Labels"))]
+    LabelBuild {
+        source: stackable_operator::kvp::LabelError,
+    },
+
+    #[snafu(display("failed to build listener volume"))]
+    BuildListenerVolume {
+        source: ListenerOperatorVolumeSourceBuilderError,
+    },
+
+    #[snafu(display("failed to apply group listener"))]
+    ApplyGroupListener {
+        source: stackable_operator::cluster_resources::Error,
+    },
 }
 
 type Result<T, E = Error> = std::result::Result<T, E>;
@@ -450,16 +468,6 @@ pub async fn reconcile_airflow(
                 role: role_name.to_string(),
             })?;
 
-        // some roles will only run "internally" and do not need to be created as services
-        if let Some(resolved_port) = role_port(role_name) {
-            let role_service =
-                build_role_service(airflow, &resolved_product_image, role_name, resolved_port)?;
-            cluster_resources
-                .add(client, role_service)
-                .await
-                .context(ApplyRoleServiceSnafu)?;
-        }
-
         for (rolegroup_name, rolegroup_config) in role_config.iter() {
             let rolegroup = RoleGroupRef {
                 cluster: ObjectRef::from_obj(airflow),
@@ -470,6 +478,18 @@ pub async fn reconcile_airflow(
             let merged_airflow_config = airflow
                 .merged_config(&airflow_role, &rolegroup)
                 .context(FailedToResolveConfigSnafu)?;
+
+            let git_sync_resources = git_sync::v1alpha1::GitSyncResources::new(
+                &airflow.spec.cluster_config.dags_git_sync,
+                &resolved_product_image,
+                &env_vars_from_rolegroup_config(rolegroup_config),
+                &airflow.volume_mounts(),
+                LOG_VOLUME_NAME,
+                &merged_airflow_config
+                    .logging
+                    .for_container(&Container::GitSync),
+            )
+            .context(InvalidGitSyncSpecSnafu)?;
 
             let rg_service = build_rolegroup_service(airflow, &resolved_product_image, &rolegroup)?;
             cluster_resources.add(client, rg_service).await.context(
@@ -489,7 +509,28 @@ pub async fn reconcile_airflow(
                 &rbac_sa,
                 &merged_airflow_config,
                 airflow_executor,
+                &git_sync_resources,
             )?;
+
+            if let Some(listener_class) =
+                airflow.merged_listener_class(&airflow_role, &rolegroup.role_group)
+            {
+                if let Some(listener_group_name) =
+                    airflow.group_listener_name(&airflow_role, &rolegroup)
+                {
+                    let rg_group_listener = build_group_listener(
+                        airflow,
+                        &resolved_product_image,
+                        &rolegroup,
+                        listener_class.to_string(),
+                        listener_group_name,
+                    )?;
+                    cluster_resources
+                        .add(client, rg_group_listener)
+                        .await
+                        .context(ApplyGroupListenerSnafu)?;
+                }
+            }
 
             ss_cond_builder.add(
                 cluster_resources
@@ -523,7 +564,7 @@ pub async fn reconcile_airflow(
             pod_disruption_budget: pdb,
         }) = role_config
         {
-            add_pdbs(pdb, airflow, &airflow_role, client, &mut cluster_resources)
+            add_pdbs(&pdb, airflow, &airflow_role, client, &mut cluster_resources)
                 .await
                 .context(FailedToCreatePdbSnafu)?;
         }
@@ -585,6 +626,19 @@ async fn build_executor_template(
         .with_context(|_| ApplyRoleGroupConfigSnafu {
             rolegroup: rolegroup.clone(),
         })?;
+
+    let git_sync_resources = git_sync::v1alpha1::GitSyncResources::new(
+        &airflow.spec.cluster_config.dags_git_sync,
+        resolved_product_image,
+        &env_vars_from(&common_config.env_overrides),
+        &airflow.volume_mounts(),
+        LOG_VOLUME_NAME,
+        &merged_executor_config
+            .logging
+            .for_container(&Container::GitSync),
+    )
+    .context(InvalidGitSyncSpecSnafu)?;
+
     let worker_pod_template_config_map = build_executor_template_config_map(
         airflow,
         resolved_product_image,
@@ -594,74 +648,13 @@ async fn build_executor_template(
         &common_config.env_overrides,
         &common_config.pod_overrides,
         &rolegroup,
+        &git_sync_resources,
     )?;
     cluster_resources
         .add(client, worker_pod_template_config_map)
         .await
         .with_context(|_| ApplyExecutorTemplateConfigSnafu {})?;
     Ok(())
-}
-
-/// The server-role service is the primary endpoint that should be used by clients that do not perform internal load balancing,
-/// including targets outside the cluster.
-fn build_role_service(
-    airflow: &v1alpha1::AirflowCluster,
-    resolved_product_image: &ResolvedProductImage,
-    role_name: &str,
-    port: u16,
-) -> Result<Service> {
-    let role_svc_name = format!("{}-{}", airflow.name_any(), role_name);
-    let ports = role_ports(port);
-
-    let metadata = ObjectMetaBuilder::new()
-        .name_and_namespace(airflow)
-        .name(&role_svc_name)
-        .ownerreference_from_resource(airflow, None, Some(true))
-        .context(ObjectMissingMetadataForOwnerRefSnafu)?
-        .with_recommended_labels(build_recommended_labels(
-            airflow,
-            AIRFLOW_CONTROLLER_NAME,
-            &resolved_product_image.app_version_label,
-            role_name,
-            "global",
-        ))
-        .context(ObjectMetaSnafu)?
-        .build();
-
-    let service_selector_labels =
-        Labels::role_selector(airflow, APP_NAME, role_name).context(BuildLabelSnafu)?;
-
-    let service_spec = ServiceSpec {
-        type_: Some(
-            airflow
-                .spec
-                .cluster_config
-                .listener_class
-                .k8s_service_type(),
-        ),
-        ports: Some(ports),
-        selector: Some(service_selector_labels.into()),
-        ..ServiceSpec::default()
-    };
-
-    Ok(Service {
-        metadata,
-        spec: Some(service_spec),
-        status: None,
-    })
-}
-
-fn role_ports(port: u16) -> Vec<ServicePort> {
-    vec![ServicePort {
-        name: Some("http".to_string()),
-        port: port.into(),
-        protocol: Some("TCP".to_string()),
-        ..ServicePort::default()
-    }]
-}
-
-fn role_port(role_name: &str) -> Option<u16> {
-    AirflowRole::from_str(role_name).unwrap().get_http_port()
 }
 
 /// The rolegroup [`ConfigMap`] configures the rolegroup based on the configuration given by the administrator
@@ -781,16 +774,12 @@ fn build_rolegroup_service(
     resolved_product_image: &ResolvedProductImage,
     rolegroup: &RoleGroupRef<v1alpha1::AirflowCluster>,
 ) -> Result<Service> {
-    let mut ports = vec![ServicePort {
+    let ports = vec![ServicePort {
         name: Some(METRICS_PORT_NAME.into()),
-        port: METRICS_PORT,
+        port: METRICS_PORT.into(),
         protocol: Some("TCP".to_string()),
         ..Default::default()
     }];
-
-    if let Some(http_port) = role_port(&rolegroup.role) {
-        ports.append(&mut role_ports(http_port));
-    }
 
     let prometheus_label =
         Label::try_from(("prometheus.io/scrape", "true")).context(BuildLabelSnafu)?;
@@ -800,6 +789,7 @@ fn build_rolegroup_service(
         &resolved_product_image,
         &rolegroup,
         prometheus_label,
+        format!("{name}-metrics", name = rolegroup.object_name()),
     )?;
 
     let service_selector_labels =
@@ -828,10 +818,11 @@ fn build_rolegroup_metadata(
     resolved_product_image: &&ResolvedProductImage,
     rolegroup: &&RoleGroupRef<v1alpha1::AirflowCluster>,
     prometheus_label: Label,
+    name: String,
 ) -> Result<ObjectMeta, Error> {
     let metadata = ObjectMetaBuilder::new()
         .name_and_namespace(airflow)
-        .name(rolegroup.object_name())
+        .name(name)
         .ownerreference_from_resource(airflow, None, Some(true))
         .context(ObjectMissingMetadataForOwnerRefSnafu)?
         .with_recommended_labels(build_recommended_labels(
@@ -845,6 +836,47 @@ fn build_rolegroup_metadata(
         .with_label(prometheus_label)
         .build();
     Ok(metadata)
+}
+
+pub fn build_group_listener(
+    airflow: &v1alpha1::AirflowCluster,
+    resolved_product_image: &ResolvedProductImage,
+    rolegroup: &RoleGroupRef<v1alpha1::AirflowCluster>,
+    listener_class: String,
+    listener_group_name: String,
+) -> Result<listener::v1alpha1::Listener> {
+    Ok(listener::v1alpha1::Listener {
+        metadata: ObjectMetaBuilder::new()
+            .name_and_namespace(airflow)
+            .name(listener_group_name)
+            .ownerreference_from_resource(airflow, None, Some(true))
+            .context(ObjectMissingMetadataForOwnerRefSnafu)?
+            .with_recommended_labels(build_recommended_labels(
+                airflow,
+                AIRFLOW_CONTROLLER_NAME,
+                &resolved_product_image.app_version_label,
+                &rolegroup.role,
+                &rolegroup.role_group,
+            ))
+            .context(ObjectMetaSnafu)?
+            .build(),
+        spec: listener::v1alpha1::ListenerSpec {
+            class_name: Some(listener_class),
+            ports: Some(listener_ports()),
+            ..listener::v1alpha1::ListenerSpec::default()
+        },
+        status: None,
+    })
+}
+
+/// We only use the http port here and intentionally omit
+/// the metrics one.
+fn listener_ports() -> Vec<listener::v1alpha1::ListenerPort> {
+    vec![listener::v1alpha1::ListenerPort {
+        name: HTTP_PORT_NAME.to_string(),
+        port: HTTP_PORT.into(),
+        protocol: Some("TCP".to_string()),
+    }]
 }
 
 /// The rolegroup [`StatefulSet`] runs the rolegroup, as configured by the administrator.
@@ -862,6 +894,7 @@ fn build_server_rolegroup_statefulset(
     service_account: &ServiceAccount,
     merged_airflow_config: &AirflowConfig,
     executor: &AirflowExecutor,
+    git_sync_resources: &git_sync::v1alpha1::GitSyncResources,
 ) -> Result<StatefulSet> {
     let binding = airflow.get_role(airflow_role);
     let role = binding.as_ref().context(NoAirflowRoleSnafu)?;
@@ -869,16 +902,34 @@ fn build_server_rolegroup_statefulset(
     let rolegroup = role.role_groups.get(&rolegroup_ref.role_group);
 
     let mut pb = PodBuilder::new();
+    let recommended_object_labels = build_recommended_labels(
+        airflow,
+        AIRFLOW_CONTROLLER_NAME,
+        &resolved_product_image.app_version_label,
+        &rolegroup_ref.role,
+        &rolegroup_ref.role_group,
+    );
+    // Used for PVC templates that cannot be modified once they are deployed
+    let unversioned_recommended_labels = Labels::recommended(build_recommended_labels(
+        airflow,
+        AIRFLOW_CONTROLLER_NAME,
+        // A version value is required, and we do want to use the "recommended" format for the other desired labels
+        "none",
+        &rolegroup_ref.role,
+        &rolegroup_ref.role_group,
+    ))
+    .context(LabelBuildSnafu)?;
 
     let pb_metadata = ObjectMetaBuilder::new()
-        .with_recommended_labels(build_recommended_labels(
-            airflow,
-            AIRFLOW_CONTROLLER_NAME,
-            &resolved_product_image.app_version_label,
-            &rolegroup_ref.role,
-            &rolegroup_ref.role_group,
-        ))
+        .with_recommended_labels(recommended_object_labels)
         .context(ObjectMetaSnafu)?
+        .with_annotation(
+            Annotation::try_from((
+                "kubectl.kubernetes.io/default-container",
+                format!("{}", Container::Airflow),
+            ))
+            .unwrap(),
+        )
         .build();
 
     pb.metadata(pb_metadata)
@@ -929,6 +980,7 @@ fn build_server_rolegroup_statefulset(
             executor,
             authentication_config,
             authorization_config,
+            git_sync_resources,
         )
         .context(BuildStatefulsetEnvVarsSnafu)?,
     );
@@ -953,10 +1005,11 @@ fn build_server_rolegroup_statefulset(
             .context(AddVolumeMountSnafu)?;
     }
 
-    if let Some(resolved_port) = airflow_role.get_http_port() {
+    // for roles with an http endpoint
+    if let Some(http_port) = airflow_role.get_http_port() {
         let probe = Probe {
             tcp_socket: Some(TCPSocketAction {
-                port: IntOrString::Int(resolved_port.into()),
+                port: IntOrString::Int(http_port.into()),
                 ..TCPSocketAction::default()
             }),
             initial_delay_seconds: Some(60),
@@ -966,8 +1019,43 @@ fn build_server_rolegroup_statefulset(
         };
         airflow_container.readiness_probe(probe.clone());
         airflow_container.liveness_probe(probe);
-        airflow_container.add_container_port("http", resolved_port.into());
+        airflow_container.add_container_port(HTTP_PORT_NAME, http_port.into());
     }
+
+    let mut pvcs: Option<Vec<PersistentVolumeClaim>> = None;
+
+    if let Some(listener_group_name) = airflow.group_listener_name(airflow_role, rolegroup_ref) {
+        // Listener endpoints for the Webserver role will use persistent volumes
+        // so that load balancers can hard-code the target addresses. This will
+        // be the case even when no class is set (and the value defaults to
+        // cluster-internal) as the address should still be consistent.
+        let pvc = ListenerOperatorVolumeSourceBuilder::new(
+            &ListenerReference::ListenerName(listener_group_name),
+            &unversioned_recommended_labels,
+        )
+        .context(BuildListenerVolumeSnafu)?
+        .build_pvc(LISTENER_VOLUME_NAME.to_string())
+        .context(BuildListenerVolumeSnafu)?;
+        pvcs = Some(vec![pvc]);
+
+        airflow_container
+            .add_volume_mount(LISTENER_VOLUME_NAME, LISTENER_VOLUME_DIR)
+            .context(AddVolumeMountSnafu)?;
+    }
+
+    // If the DAG is modularized we may encounter a timing issue whereby the celery worker
+    // has started *before* all modules referenced by the DAG have been fetched by gitsync
+    // and registered. This will result in ModuleNotFoundError errors. This can be avoided
+    // by running a one-off git-sync process in an init-container so that all DAG
+    // dependencies are fully loaded. The sidecar git-sync is then used for regular updates.
+    let use_git_sync_init_containers = matches!(executor, AirflowExecutor::CeleryExecutor { .. });
+    add_git_sync_resources(
+        &mut pb,
+        &mut airflow_container,
+        git_sync_resources,
+        true,
+        use_git_sync_init_containers,
+    )?;
 
     pb.add_container(airflow_container.build());
 
@@ -990,7 +1078,7 @@ fn build_server_rolegroup_statefulset(
             ]
             .join("\n"),
         ])
-        .add_container_port(METRICS_PORT_NAME, METRICS_PORT)
+        .add_container_port(METRICS_PORT_NAME, METRICS_PORT.into())
         .resources(
             ResourceRequirementsBuilder::new()
                 .with_cpu_request("100m")
@@ -1020,43 +1108,6 @@ fn build_server_rolegroup_statefulset(
                 .build(),
         )
         .context(AddVolumeSnafu)?;
-    }
-
-    if let Some(gitsync) = airflow.git_sync() {
-        let gitsync_container = build_gitsync_container(
-            resolved_product_image,
-            &gitsync,
-            false,
-            &format!("{}-{}", GIT_SYNC_NAME, 1),
-            build_gitsync_statefulset_envs(rolegroup_config, &gitsync.credentials_secret),
-            airflow.volume_mounts(),
-        )?;
-
-        pb.add_volume(
-            VolumeBuilder::new(GIT_SYNC_CONTENT)
-                .empty_dir(EmptyDirVolumeSource::default())
-                .build(),
-        )
-        .context(AddVolumeSnafu)?;
-
-        pb.add_container(gitsync_container);
-
-        if let AirflowExecutor::CeleryExecutor { .. } = executor {
-            let gitsync_init_container = build_gitsync_container(
-                resolved_product_image,
-                &gitsync,
-                true,
-                &format!("{}-{}", GIT_SYNC_NAME, 0),
-                build_gitsync_statefulset_envs(rolegroup_config, &gitsync.credentials_secret),
-                airflow.volume_mounts(),
-            )?;
-            // If the DAG is modularized we may encounter a timing issue whereby the celery worker has started
-            // *before* all modules referenced by the DAG have been fetched by gitsync and registered. This
-            // will result in ModuleNotFoundError errors. This can be avoided by running a one-off git-sync
-            // process in an init-container so that all DAG dependencies are fully loaded. The sidecar
-            // git-sync is then used for regular updates.
-            pb.add_init_container(gitsync_init_container);
-        }
     }
 
     if merged_airflow_config.logging.enable_vector_agent {
@@ -1094,6 +1145,7 @@ fn build_server_rolegroup_statefulset(
         &resolved_product_image,
         &rolegroup_ref,
         restarter_label,
+        rolegroup_ref.object_name(),
     )?;
 
     let statefulset_match_labels = Labels::role_group_selector(
@@ -1119,8 +1171,9 @@ fn build_server_rolegroup_statefulset(
             match_labels: Some(statefulset_match_labels.into()),
             ..LabelSelector::default()
         },
-        service_name: rolegroup_ref.object_name(),
+        service_name: Some(rolegroup_ref.object_name()),
         template: pod_template,
+        volume_claim_templates: pvcs,
         ..StatefulSetSpec::default()
     };
 
@@ -1162,6 +1215,7 @@ fn build_executor_template_config_map(
     env_overrides: &HashMap<String, String>,
     pod_overrides: &PodTemplateSpec,
     rolegroup_ref: &RoleGroupRef<v1alpha1::AirflowCluster>,
+    git_sync_resources: &git_sync::v1alpha1::GitSyncResources,
 ) -> Result<ConfigMap> {
     let mut pb = PodBuilder::new();
     let pb_metadata = ObjectMetaBuilder::new()
@@ -1205,10 +1259,12 @@ fn build_executor_template_config_map(
     airflow_container
         .image_from_product_image(resolved_product_image)
         .resources(merged_executor_config.resources.clone().into())
-        .add_env_vars(
-            build_airflow_template_envs(airflow, env_overrides, merged_executor_config)
-                .context(BuildStatefulsetEnvVarsSnafu)?,
-        )
+        .add_env_vars(build_airflow_template_envs(
+            airflow,
+            env_overrides,
+            merged_executor_config,
+            git_sync_resources,
+        ))
         .add_volume_mounts(airflow.volume_mounts())
         .context(AddVolumeMountSnafu)?
         .add_volume_mount(CONFIG_VOLUME_NAME, CONFIG_PATH)
@@ -1217,6 +1273,14 @@ fn build_executor_template_config_map(
         .context(AddVolumeMountSnafu)?
         .add_volume_mount(LOG_VOLUME_NAME, STACKABLE_LOG_DIR)
         .context(AddVolumeMountSnafu)?;
+
+    add_git_sync_resources(
+        &mut pb,
+        &mut airflow_container,
+        git_sync_resources,
+        false,
+        true,
+    )?;
 
     pb.add_container(airflow_container.build());
     pb.add_volumes(airflow.volumes().clone())
@@ -1229,24 +1293,6 @@ fn build_executor_template_config_map(
             .get(&Container::Airflow),
     ))
     .context(AddVolumeSnafu)?;
-
-    if let Some(gitsync) = airflow.git_sync() {
-        let gitsync_container = build_gitsync_container(
-            resolved_product_image,
-            &gitsync,
-            true,
-            &format!("{}-{}", GIT_SYNC_NAME, 0),
-            build_gitsync_template(env_overrides, &gitsync.credentials_secret),
-            airflow.volume_mounts(),
-        )?;
-        pb.add_volume(
-            VolumeBuilder::new(GIT_SYNC_CONTENT)
-                .empty_dir(EmptyDirVolumeSource::default())
-                .build(),
-        )
-        .context(AddVolumeSnafu)?;
-        pb.add_init_container(gitsync_container);
-    }
 
     if merged_executor_config.logging.enable_vector_agent {
         match &airflow
@@ -1304,42 +1350,6 @@ fn build_executor_template_config_map(
     cm_builder.build().context(PodTemplateConfigMapSnafu)
 }
 
-fn build_gitsync_container(
-    resolved_product_image: &ResolvedProductImage,
-    gitsync: &&GitSync,
-    one_time: bool,
-    name: &str,
-    env_vars: Vec<EnvVar>,
-    volume_mounts: Vec<VolumeMount>,
-) -> Result<k8s_openapi::api::core::v1::Container, Error> {
-    let gitsync_container = ContainerBuilder::new(name)
-        .context(InvalidContainerNameSnafu)?
-        .add_env_vars(env_vars)
-        .image_from_product_image(resolved_product_image)
-        .command(vec![
-            "/bin/bash".to_string(),
-            "-x".to_string(),
-            "-euo".to_string(),
-            "pipefail".to_string(),
-            "-c".to_string(),
-        ])
-        .args(vec![gitsync.get_args(one_time).join("\n")])
-        .add_volume_mount(GIT_SYNC_CONTENT, GIT_SYNC_ROOT)
-        .context(AddVolumeMountSnafu)?
-        .add_volume_mounts(volume_mounts)
-        .context(AddVolumeMountSnafu)?
-        .resources(
-            ResourceRequirementsBuilder::new()
-                .with_cpu_request("100m")
-                .with_cpu_limit("200m")
-                .with_memory_request("64Mi")
-                .with_memory_limit("64Mi")
-                .build(),
-        )
-        .build();
-    Ok(gitsync_container)
-}
-
 pub fn error_policy(
     _obj: Arc<DeserializeGuard<v1alpha1::AirflowCluster>>,
     error: &Error,
@@ -1387,5 +1397,30 @@ fn add_authentication_volumes_and_volume_mounts(
         tls.add_volumes_and_mounts(pb, vec![cb])
             .context(AddTlsVolumesAndVolumeMountsSnafu)?;
     }
+    Ok(())
+}
+
+fn add_git_sync_resources(
+    pb: &mut PodBuilder,
+    cb: &mut ContainerBuilder,
+    git_sync_resources: &git_sync::v1alpha1::GitSyncResources,
+    add_sidecar_containers: bool,
+    add_init_containers: bool,
+) -> Result<()> {
+    if add_sidecar_containers {
+        for container in git_sync_resources.git_sync_containers.iter().cloned() {
+            pb.add_container(container);
+        }
+    }
+    if add_init_containers {
+        for container in git_sync_resources.git_sync_init_containers.iter().cloned() {
+            pb.add_init_container(container);
+        }
+    }
+    pb.add_volumes(git_sync_resources.git_content_volumes.to_owned())
+        .context(AddVolumeSnafu)?;
+    cb.add_volume_mounts(git_sync_resources.git_content_volume_mounts.to_owned())
+        .context(AddVolumeMountSnafu)?;
+
     Ok(())
 }