+++ tests/templates/kuttl/commons/metrics.py
@@ -7,20 +7,25 @@ from datetime import datetime, timezone
 import argparse
 import logging
 
+
 def exception_handler(exception_type, exception, traceback):
     print(f"{exception_type.__name__}: {exception.args}")
 
+
 sys.excepthook = exception_handler
 
+
 def assert_metric(role, role_group, metric):
-    metric_response = requests.get(f"http://airflow-{role}-{role_group}:9102/metrics")
+    metric_response = requests.get(
+        f"http://airflow-{role}-{role_group}-metrics:9102/metrics"
+    )
     assert metric_response.status_code == 200, (
         f"Metrics could not be retrieved from the {role}-{role_group}."
     )
     return metric in metric_response.text
 
-def metrics_v3(role_group: str) -> None:
 
+def metrics_v3(role_group: str) -> None:
     now = datetime.now(timezone.utc)
     ts = now.strftime("%Y-%m-%dT%H:%M:%S.%f") + now.strftime("%z")
 
@@ -47,7 +52,9 @@ def metrics_v3(role_group: str) -> None:
         access_token = token_data["access_token"]
         print(f"Access Token: {access_token}")
     else:
-        print(f"Failed to obtain access token: {response.status_code} - {response.text}")
+        print(
+            f"Failed to obtain access token: {response.status_code} - {response.text}"
+        )
         sys.exit(1)
 
     headers = {
@@ -77,7 +84,9 @@ def metrics_v3(role_group: str) -> None:
             (assert_metric("scheduler", role_group, "airflow_scheduler_heartbeat"))
             and (
                 assert_metric(
-                    "webserver", role_group, "airflow_task_instance_created_BashOperator"
+                    "webserver",
+                    role_group,
+                    "airflow_task_instance_created_BashOperator",
                 )
             )  # noqa: W503, W504
             and (
@@ -95,25 +104,25 @@ def metrics_v3(role_group: str) -> None:
             # force re-try of script
             sys.exit(1)
 
+
 def metrics_v2(role_group: str) -> None:
-    
     # Trigger a DAG run to create metrics
     dag_id = "example_trigger_target_dag"
     dag_conf = {"message": "Hello World"}
-    
+
     rest_url = f"http://airflow-webserver-{role_group}:8080/api/v1"
     auth = ("airflow", "airflow")
-    
+
     # allow a few moments for the DAGs to be registered to all roles
     time.sleep(10)
-    
+
     response = requests.patch(
         f"{rest_url}/dags/{dag_id}", auth=auth, json={"is_paused": False}
     )
     response = requests.post(
         f"{rest_url}/dags/{dag_id}/dagRuns", auth=auth, json={"conf": dag_conf}
     )
-    
+
     # Test the DAG in a loop. Each time we call the script a new job will be started: we can avoid
     # or minimize this by looping over the check instead.
     iterations = 4
@@ -127,7 +136,9 @@ def metrics_v2(role_group: str) -> None:
             (assert_metric("scheduler", role_group, "airflow_scheduler_heartbeat"))
             and (
                 assert_metric(
-                    "webserver", role_group, "airflow_task_instance_created_BashOperator"
+                    "webserver",
+                    role_group,
+                    "airflow_task_instance_created_BashOperator",
                 )
             )  # noqa: W503, W504
             and (
@@ -145,8 +156,8 @@ def metrics_v2(role_group: str) -> None:
             # force re-try of script
             sys.exit(1)
 
-if __name__ == "__main__":
 
+if __name__ == "__main__":
     log_level = "DEBUG"
     logging.basicConfig(
         level=log_level,
@@ -155,7 +166,9 @@ if __name__ == "__main__":
     )
 
     parser = argparse.ArgumentParser(description="Airflow metrics script")
-    parser.add_argument("--role-group", type=str, default="default", help="Role group to check")
+    parser.add_argument(
+        "--role-group", type=str, default="default", help="Role group to check"
+    )
     parser.add_argument("--airflow-version", type=str, help="Airflow version")
     opts = parser.parse_args()